requirements.txt
async-generator==1.10
attrs==21.2.0
autopep8==1.5.7
beautifulsoup4==4.10.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
cryptography==35.0.0
h11==0.12.0
idna==3.2
numpy==1.21.2
outcome==1.1.0
pandas==1.3.3
pycodestyle==2.7.0
pycparser==2.20
PyMySQL==1.0.2
pyOpenSSL==21.0.0
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
selenium==4.0.0b4
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.2.1
toml==0.10.2
trio==0.19.0
trio-websocket==0.9.2
undetected-chromedriver==3.0.3
urllib3==1.26.7
websockets==10.0
wsproto==1.0.0
